% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ruban.R
\name{nuclearFunction}
\alias{nuclearFunction}
\title{Use of nuclear function}
\usage{
nuclearFunction(z, x = 1, r = 2, s = 3)
}
\arguments{
\item{z}{value}

\item{x}{choice of nuclear function}

\item{r}{nuclear function parameter}

\item{s}{nuclear function parameter}
}
\value{
result of nuclear function
}
\description{
The choice is made at the discretion of the user
}
\examples{
z <- 5
x <- nuclearFunction(z)
print(z)
}
