% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rouban.R
\docType{package}
\name{Rouban}
\alias{Rouban}
\alias{Rouban-package}
\title{Rouban: package for finding global extrema global optimization method based on the selective averaging coordinate  with restrictions.}
\usage{
Rouban(x, delta, fitness, lower, upper, n = 500, e = 0.001, M = 1000,
  y = 1, q = 2, nuclearType = "nuclearExponential", r = 2,
  s = 100)
}
\arguments{
\item{x}{starting point coordinate}

\item{delta}{increment x denotes search range}

\item{fitness}{function to search for extremum}

\item{lower}{lower extremum limits}

\item{upper}{upper extremum limits}

\item{n}{amount of test points}

\item{e}{precision constant}

\item{M}{maximum number of iterations}

\item{y}{stretch factor}

\item{q}{matched fixed parameter}

\item{nuclearType}{type of nuclear function}

\item{r}{matched fixed parameter}

\item{s}{core selectivity factor}
}
\value{
potential point extremum
}
\description{
Rouban: package for finding global extrema global optimization method based on the selective averaging coordinate  with restrictions.

Search for extremum value using global optimization method based on the selective averaging coordinate  with restrictions.
}
\examples{
f<-function(x) {
z<-7*(abs(x[1])^2) + 7*(abs(x[2])^2)
z<-c(z, 5*(abs(x[1]-3))^0.8 + 5*(abs(x[2]-3)^0.6) + 6)
z<-c(z, 5*(abs(x[1]-6))^1.3 + 5*(abs(x[2]-6)^1.3) + 2)
z<-c(z, 5*(abs(x[1]-6))^1 + 5*(abs(x[2]+6)^1) + 8)
z<-c(z, 4*(abs(x[1]+6))^1.5 + 4*(abs(x[2]+6)^1.5) + 7)
z<-c(z, 5*(abs(x[1]+3))^1.8 + 5*(abs(x[2])^1.8) + 9)
z<-c(z, 6*(abs(x[1]+6))^0.6 + 6*(abs(x[2]-6)^0.9))
return(min(z))
}
x<-Rouban(x=c(10,10),
delta=c(20,20),
lower=c(-10,-10),
upper = c(30,30),
fitness = f,
n=500,
y=1,
q=2,
s=100,
e=0.0001,
r=2,
nuclearType = "nuclearExponential")
print(x)
}
