% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ruban.R
\docType{package}
\name{Ruban}
\alias{Ruban}
\alias{Ruban-package}
\title{Ruban: package allows you to use the algorithm for finding global extrema by averaging coordinates}
\usage{
Ruban(x, deltaX, f, lower, upper, n = 500, e = 0.001, M = 1000,
  y = 1, q = 2, nuclear_function = 1, r = 2, s = 100)
}
\arguments{
\item{x}{starting point coordinate}

\item{deltaX}{increment x denotes search range}

\item{f}{search function}

\item{lower}{lower extremum limits}

\item{upper}{upper extremum limits}

\item{n}{amount of points}

\item{e}{precision constant}

\item{M}{maximum number of iterations}

\item{y}{parameter of Ruban's algorithm}

\item{q}{parameter of Ruban's algorithm}

\item{nuclear_function}{choice of nuclear function}

\item{r}{parameter for nuclear function}

\item{s}{parameter for nuclear function}
}
\value{
numeric the estimated amount of good honey
}
\description{
Ruban: package allows you to use the algorithm for finding global extrema by averaging coordinates

Search for extremum value using coordinate averaging method
}
\examples{
f<-function(x) {
z<-7*(abs(x[1])^2) + 7*(abs(x[2])^2)
z<-c(z, 5*(abs(x[1]-3))^0.8 + 5*(abs(x[2]-3)^0.6) + 6)
z<-c(z, 5*(abs(x[1]-6))^1.3 + 5*(abs(x[2]-6)^1.3) + 2)
z<-c(z, 5*(abs(x[1]-6))^1 + 5*(abs(x[2]+6)^1) + 8)
z<-c(z, 4*(abs(x[1]+6))^1.5 + 4*(abs(x[2]+6)^1.5) + 7)
z<-c(z, 5*(abs(x[1]+3))^1.8 + 5*(abs(x[2])^1.8) + 9)
z<-c(z, 6*(abs(x[1]+6))^0.6 + 6*(abs(x[2]-6)^0.9))
return(min(z))
}
x<-Ruban(x=c(10,10),
deltaX=c(20,20),
lower=c(-10,-10),
upper = c(30,30),
f = f,
n=500,
y=1,
q=2,
s=100,
e=0.0001,
r=2,
nuclear_function = 1)
print(x)
}
